[CTRLS-78] <feature> 첫 번째 페이즈 패턴 구현 및 일부 효과음 추가
1페이즈에서 2페이즈 넘어가는 경우의 기본적인 패턴을 구현하였습니다. 이후에 수정될 수도 있습니다.
패턴에 사용하기 위해 화면상에 가로와 세로로 총알을 일자로 출력하는 메소드를 추가하였습니다.
보스 총알의 속도를 좀 더 다양하게 설정하기 위해 speed를 double로 바꾸었습니다.

크리스탈을 때릴 때와 총알을 발사할 때 효과음을 출력하도록 바꾸었습니다.
또한 무적인 보스를 때릴 때 다른 소리를 출력하도록 하였습니다.

	modified:   src/main/java/kr/ac/hanyang/engine/DrawManager.java
	modified:   src/main/java/kr/ac/hanyang/engine/SoundManager.java
	modified:   src/main/java/kr/ac/hanyang/entity/Bullet.java
	modified:   src/main/java/kr/ac/hanyang/entity/BulletPool.java
	modified:   src/main/java/kr/ac/hanyang/entity/boss/Boss.java
	modified:   src/main/java/kr/ac/hanyang/screen/BossScreen.java
	modified:   src/main/java/kr/ac/hanyang/screen/GameScreen.java
	modified:   src/main/resources/graphics
	new file:   src/main/resources/sounds/CrystalBreak.wav
	new file:   src/main/resources/sounds/InvicibleHit.wav
	new file:   src/main/resources/sounds/Lasergun.wav

+++ src/main/java/kr/ac/hanyang/entity/BulletPool.java
@@ -65,7 +65,7 @@ public final class BulletPool {
     }
     //보스 총알인 경우의 getBullet
     public static Bullet getBossBullet(final int positionX,
-        final int positionY, final int speed, int damage, double angle) {
+        final int positionY, final double speed, int damage, double angle) {
         Bullet bullet;
         if (!pool.isEmpty()) {
             bullet = pool.iterator().next();
